segfaults when it gets to a paragraph that ends with a space
The uptake of text works well. I haven't pasted in the parsing from unputtest that looks at command line, but I've been working on this for hours... the segfault occurs when it hits a space before a paragraph at the end. it doesn't print the last line. Also it doesn't seem to be checking the width very well.

Make sure that you use a robust testfile with this.